plat: Introduce `lcpu_get_current_in_except()`
Assuming we are in an exception handler on an exception stack,
fetch  the current LCPU pointer in such a manner that we do not
assume any register state. This is done by dividng current stack
pointer relative to the LCPU stack base to find out the current
LCPU id.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1688