lib/ukfs: Update to typed vopen API
This change updates the ukfs driver API with a vopen() that accepts
typed arguments for volume and data. This allows drivers to request
preprocessing of arguments by the caller without needing to "call up"
the abstraction stack, and adding undue dependencies.
This is a breaking change and existing drivers and callers are updated.

Checkpatch-Ignore: SPACING
Checkpatch-Ignore: MACRO_ARG_REUSE
Signed-off-by: Andrei Tatar <ttr@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1724