lib/posix-fstab: Correctly handle mounting initrd
Previously when mounting (not extracting) an initrd, fstab would call
mount() with the initrd base as source, leaving posix-vfs or the driver
to correctly interpret it as a memory image.
With the ukfs API update to typed vopen, passing memory images must be
done explicitly; this change update fstab to correctly pass the initrd
to vopen before performing the VFS mount.

Signed-off-by: Andrei Tatar <ttr@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1724