lib/ukrandom: Introduce CSPRNG reseed
Add uk_random_reseed() into the libukrandom API. The intention of this
function is to allow the system to reseed the CSPRNG on events like
resume. This function returns an error if there is no underlying driver.

Update random number generation to reseed the CSPRNG when the counter
reaches its maximum value. If the CSPRNG is not backed by a device
driver, exhaustion of the counter results into a fatal system error.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Andrei Tatar <ttr@unikraft.io>
Reviewed-by: Andrei Tatar <ttr@unikraft.io>
GitHub-Closes: #1732