lib/ukrandom: Fix incorrect size parameters
The number of bytes requested from the driver for the seed as well as
the number of bytes compared to determine whether a seed is set in the
cmdline, use CHACHA_SEED_LENGTH. That number is in words, not bytes,
resulting in insufficient seeding in the former case, and an invalid
comparison in the latter. Use buffer size instead.

Signed-off-by: Michalis Pappas <michalis@unikraft.io>
Approved-by: Andrei Tatar <ttr@unikraft.io>
Reviewed-by: Andrei Tatar <ttr@unikraft.io>
GitHub-Closes: #1731