lib/posix-process: Declare mainlike to not return
This change has uk_posix_process_run take a mainlike function that,
appropriately to the ukthread and multiprocess requirements, is declared
noreturn in line with the expectation that it call exit() instead of
returning. The one callsite of process_run has been patched to supply
this correct wrapper function around do_main.

Signed-off-by: Andrei Tatar <ttr@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Sergiu Moga <sergiu@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #1746